#2015: Moved and renamed metadata device client.
