+++ internal/core/command/router.go
@@ -21,7 +21,7 @@ import (
 	"github.com/edgexfoundry/go-mod-core-contracts/clients/logger"
 	"github.com/gorilla/mux"
 
-	container "github.com/edgexfoundry/edgex-go/internal/core/metadata/containers"
+	container "github.com/edgexfoundry/edgex-go/internal/core/command/containers"
 	"github.com/edgexfoundry/edgex-go/internal/pkg"
 	bootstrapContainer "github.com/edgexfoundry/edgex-go/internal/pkg/bootstrap/container"
 	"github.com/edgexfoundry/edgex-go/internal/pkg/correlation"
@@ -67,7 +67,7 @@ func loadDeviceRoutes(b *mux.Router, dic *di.Container) {
 			r,
 			bootstrapContainer.LoggingClientFrom(dic.Get),
 			bootstrapContainer.DBClientFrom(dic.Get),
-			container.DeviceClientFrom(dic.Get))
+			container.MetadataDeviceClientFrom(dic.Get))
 	}).Methods(http.MethodGet)
 
 	d := b.PathPrefix("/" + DEVICE).Subrouter()
@@ -79,7 +79,7 @@ func loadDeviceRoutes(b *mux.Router, dic *di.Container) {
 			r,
 			bootstrapContainer.LoggingClientFrom(dic.Get),
 			bootstrapContainer.DBClientFrom(dic.Get),
-			container.DeviceClientFrom(dic.Get))
+			container.MetadataDeviceClientFrom(dic.Get))
 	}).Methods(http.MethodGet)
 	d.HandleFunc("/{"+ID+"}/"+COMMAND+"/{"+COMMANDID+"}", func(w http.ResponseWriter, r *http.Request) {
 		restGetDeviceCommandByCommandID(
@@ -87,7 +87,7 @@ func loadDeviceRoutes(b *mux.Router, dic *di.Container) {
 			r,
 			bootstrapContainer.LoggingClientFrom(dic.Get),
 			bootstrapContainer.DBClientFrom(dic.Get),
-			container.DeviceClientFrom(dic.Get))
+			container.MetadataDeviceClientFrom(dic.Get))
 	}).Methods(http.MethodGet)
 	d.HandleFunc("/{"+ID+"}/"+COMMAND+"/{"+COMMANDID+"}", func(w http.ResponseWriter, r *http.Request) {
 		restPutDeviceCommandByCommandID(
@@ -95,7 +95,7 @@ func loadDeviceRoutes(b *mux.Router, dic *di.Container) {
 			r,
 			bootstrapContainer.LoggingClientFrom(dic.Get),
 			bootstrapContainer.DBClientFrom(dic.Get),
-			container.DeviceClientFrom(dic.Get))
+			container.MetadataDeviceClientFrom(dic.Get))
 	}).Methods(http.MethodPut)
 
 	// /api/<version>/device/name
@@ -107,7 +107,7 @@ func loadDeviceRoutes(b *mux.Router, dic *di.Container) {
 			r,
 			bootstrapContainer.LoggingClientFrom(dic.Get),
 			bootstrapContainer.DBClientFrom(dic.Get),
-			container.DeviceClientFrom(dic.Get))
+			container.MetadataDeviceClientFrom(dic.Get))
 	}).Methods(http.MethodGet)
 	dn.HandleFunc("/{"+NAME+"}/"+COMMAND+"/{"+COMMANDNAME+"}", func(w http.ResponseWriter, r *http.Request) {
 		restGetDeviceCommandByNames(
@@ -115,7 +115,7 @@ func loadDeviceRoutes(b *mux.Router, dic *di.Container) {
 			r,
 			bootstrapContainer.LoggingClientFrom(dic.Get),
 			bootstrapContainer.DBClientFrom(dic.Get),
-			container.DeviceClientFrom(dic.Get))
+			container.MetadataDeviceClientFrom(dic.Get))
 	}).Methods(http.MethodGet)
 	dn.HandleFunc("/{"+NAME+"}/"+COMMAND+"/{"+COMMANDNAME+"}", func(w http.ResponseWriter, r *http.Request) {
 		restPutDeviceCommandByNames(
@@ -123,7 +123,7 @@ func loadDeviceRoutes(b *mux.Router, dic *di.Container) {
 			r,
 			bootstrapContainer.LoggingClientFrom(dic.Get),
 			bootstrapContainer.DBClientFrom(dic.Get),
-			container.DeviceClientFrom(dic.Get))
+			container.MetadataDeviceClientFrom(dic.Get))
 	}).Methods(http.MethodPut)
 }
